requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
azure-core==1.36.0
azure-identity==1.25.1
azure-servicebus==7.14.3
azure-storage-blob==12.27.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
confluent-kafka==2.12.1
cryptography==46.0.3
exceptiongroup==1.3.0
fastapi==0.121.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
isodate==0.7.2
kafka-python==2.2.15
msal==1.34.0
msal-extensions==1.3.1
numpy==2.0.2
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
pandas==2.3.3
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
zipp==3.23.0
pytest==8.3.4
pytest-asyncio==0.25.0
python-dotenv==1.0.1
